requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0
geopandas>=0.14.0
requests>=2.28.0
shapely>=2.0.0
fiona>=1.9.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
scipy>=1.11.0
networkx>=3.0
